==> Code/5_Nobel_Prize_Table.ipynb <==
# Setting Data: Nobel Prize Winners 2017 - 2025

This notebook was used to manually input the names of all relevant (2017-2025) Nobel Prize winners and the corresponding year they were awarded the Nobel Memorial Prize in Economic Sciences. This information was then merged with the dataset output from Notebook 4, finalizing the dataset for analysis.

import pandas as pd
econ = pd.read_csv("../Data/economists_monthly_pageviews_2017_2025.csv")
econ[["qid", "name"]].drop_duplicates().head()


nobel = pd.DataFrame(columns=[
    "qid",
    "name",
    "nobel_winner",
    "nobel_year"
])

def add_nobel_winner(df, qid, year):
    name = econ.loc[econ["qid"] == qid, "name"].iloc[0]
    return pd.concat([
        df,
        pd.DataFrame([{
            "qid": qid,
            "name": name,
            "nobel_winner": 1,
            "nobel_year": year
        }])
    ], ignore_index=True)

nobel = add_nobel_winner(nobel, "Q7174757", 2025) # Peter Howitt
nobel = add_nobel_winner(nobel, "Q376364", 2025) # Philippe Aghion 
nobel = add_nobel_winner(nobel, "Q2899100", 2025) # Joel Mokyr 
nobel = add_nobel_winner(nobel, "Q1461050", 2024) # Simon Johnson
nobel = add_nobel_winner(nobel, "Q718581", 2024) # Daron Acemoglu
nobel = add_nobel_winner(nobel, "Q1097475", 2023) # Claudia Goldin
nobel = add_nobel_winner(nobel, "Q24565948", 2022) # Philip H. Dybvig
nobel = add_nobel_winner(nobel, "Q201795", 2022) # Ben Bernanke
nobel = add_nobel_winner(nobel, "Q185871", 2022) # Douglas Diamond
nobel = add_nobel_winner(nobel, "Q921499", 2021) # Joshua Angrist
nobel = add_nobel_winner(nobel, "Q1173928", 2021) # David Card
nobel = add_nobel_winner(nobel, "Q1359990", 2020) # Paul Milgrom
nobel = add_nobel_winner(nobel, "Q1635006", 2019) # Michael Kremer
nobel = add_nobel_winner(nobel, "Q434509", 2019) # Esther Duflo
nobel = add_nobel_winner(nobel, "Q320578", 2019) # Abhijit Banerjee
nobel = add_nobel_winner(nobel, "Q562481", 2018) # William Nordhaus
nobel = add_nobel_winner(nobel, "Q509262", 2018) # Paul Romer
nobel = add_nobel_winner(nobel, "Q371212", 2017) # Richard Thaler

nobel


nobel.to_csv("../Data/nobel_economists.csv", index=False)
