==> Code/9_Zero_Shot_Text_Classification.ipynb <==
# Text Classification with Zero-Shot

In this notebook, I used Zero-Shot Classification on Wikipedia summary texts to categorize economists into standard subfields (e.g., Macroeconomics, Financial Economics, Public Economics) without needing a pre-labeled training dataset.

**Table of Content**

1. [Set-up & Load Data](#sec1)
2. [Load Zero Shot Classifier](#sec2)
3. [Classify Summaries](#sec3)
4. [Save Results](#sec4)
5. [Accuracy Check](#sec5)
6. [Exporting Results](#sec6)

<a id="sec1"></a>
### Set-up & Load Data

For the classification process, we only need the summaries of the economists.

import pandas as pd
from transformers import pipeline
from tqdm import tqdm

df = pd.read_csv("../Data/economists_final_dataset.csv")

df = df[["qid", "name", "summary"]].dropna(subset=["summary"])
df.head()

Previously, many rows got duplicated multiple times. The code below ensures that only unique rows are stored

# Total Rows 
print("Total rows:", len(df))
# Unique Rows
print(
    "Unique qid-date pairs:",
    df[["qid", "name"]].drop_duplicates().shape[0]
)
# Duplicated Rows
dupes = df.duplicated(subset=["qid", "name"]).sum()
print("Duplicate qid-date rows:", dupes)

df = df.drop_duplicates(subset=["qid", "name"])
df.head()

<a id="sec2"></a>
### Load Zero Shot Classifier

classifier = pipeline(
    "zero-shot-classification",
    model="facebook/bart-large-mnli"
)

econ_categories = [
    "macroeconomics",
    "microeconomics",
    "development economics",
    "financial economics",
    "labor economics",
    "political economy",
    "economic history",
    "econometrics",
    "public policy",
    "other"
]

<a id="sec3"></a>
### Classify Summaries

results = []

for _, row in tqdm(df.iterrows(), total=len(df)):
    out = classifier(
        row["summary"][:512],
        candidate_labels = econ_categories
    )
    
    results.append({
        "qid": row["qid"],
        "econ_field_pred": out["labels"][0],
        "econ_field_score": out["scores"][0]
    })


<a id="sec4"></a>
### Save Results

field_df = pd.DataFrame(results)
field_df.to_csv("../Data/economist_text_classification.csv", index=False)

<a id="sec5"></a>
### Accuracy Check

ground_truth = {
    "Q718581": "macroeconomics",  # Acemoglu
    "Q272731": "public policy",  # Edith Abbott
    "Q160270": "trade", # Ricardo
    "Q9061": "sociology", # Marx
    "Q13529" : "neoclassical",
    "Q9317": "macroeconomics",
    "Q9381": "classical",
    "Q1325": "liberalism",
    "Q102454": "theory",
    "Q192592": "equilibrium",
    "Q191020": "microeconomics",
    "Q132489": "welfare",
    "Q153761": "governance",
    "Q233950": "behavioral",
    "Q131112": "geography",
    "Q434509": "development",
    "Q1097475": "labor",
    "Q562481": "environment",
    "Q263725": "policy"
}

eval_df = field_df[field_df["qid"].isin(ground_truth.keys())].copy()
eval_df["true_label"] = eval_df["qid"].map(ground_truth)
eval_df["correct"] = eval_df["econ_field_pred"] == eval_df["true_label"]

eval_df["correct"].mean()

Because economists often span multiple subfields, strict accuracy evaluation is difficult.

<a id="sec6"></a>
### Exporting Results

views = pd.read_csv("../Data/nobel_combined.csv")
views = views[["qid", "name", "date", "views", "is_nobel", "nobel_year"]]
views = views.drop_duplicates(subset=["qid", "date"])

print(len(views))

meta = pd.read_csv("../Data/economists_final_dataset.csv")
meta = meta[["qid", "gender", "citizenship", "birth_year"]].drop_duplicates(subset=["qid"])

fields = pd.read_csv("../Data/economist_text_classification.csv")
fields = fields[["qid", "econ_field_pred", "econ_field_score"]].drop_duplicates(subset=["qid"])

df = (views.merge(meta, on="qid", how="left", validate="many_to_one")
      .merge(fields, on="qid", how="left", validate="many_to_one"))

print(df.shape)

df.head()

df.to_csv("../Data/final_streamlit_dataset.csv", index=False)
